backend/requirements.txt
fastapi>=0.111
uvicorn[standard]
celery[redis]
sqlalchemy[asyncio]>=2.0
asyncpg
alembic
pdfplumber
pdf2image
pytesseract
Pillow
tiktoken
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
email-validator
redis
psycopg2-binary
openai>=1.0
